IGNITE-2627: Removed OdbcNioParser. Using GridBufferedParser for message separation. Message parsing moved to OdbcNioListener.
