+++ modules/core/src/main/java/org/apache/ignite/internal/processors/odbc/OdbcNioListener.java
@@ -17,42 +17,31 @@
 
 package org.apache.ignite.internal.processors.odbc;
 
-import org.apache.ignite.IgniteCache;
 import org.apache.ignite.IgniteLogger;
-import org.apache.ignite.cache.query.QueryCursor;
-import org.apache.ignite.cache.query.SqlFieldsQuery;
 import org.apache.ignite.internal.GridKernalContext;
-import org.apache.ignite.internal.processors.cache.QueryCursorImpl;
-import org.apache.ignite.internal.processors.query.GridQueryFieldMetadata;
-import org.apache.ignite.internal.processors.query.GridQueryTypeDescriptor;
+import org.apache.ignite.internal.binary.BinaryRawWriterEx;
+import org.apache.ignite.internal.binary.BinaryReaderExImpl;
+import org.apache.ignite.internal.binary.GridBinaryMarshaller;
+import org.apache.ignite.internal.binary.streams.BinaryHeapInputStream;
+import org.apache.ignite.internal.binary.streams.BinaryHeapOutputStream;
+import org.apache.ignite.internal.binary.streams.BinaryInputStream;
+import org.apache.ignite.internal.processors.cache.binary.CacheObjectBinaryProcessorImpl;
 import org.apache.ignite.internal.util.GridSpinBusyLock;
 import org.apache.ignite.internal.util.nio.GridNioServerListenerAdapter;
 import org.apache.ignite.internal.util.nio.GridNioSession;
-import org.apache.ignite.internal.util.typedef.F;
 import org.apache.ignite.internal.util.typedef.internal.U;
-import org.apache.ignite.lang.IgniteBiTuple;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.ArrayList;
+import java.io.IOException;
+import java.util.Arrays;
 import java.util.Collection;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicLong;
-
-import static org.apache.ignite.internal.processors.odbc.OdbcRequest.CLOSE_SQL_QUERY;
-import static org.apache.ignite.internal.processors.odbc.OdbcRequest.EXECUTE_SQL_QUERY;
-import static org.apache.ignite.internal.processors.odbc.OdbcRequest.FETCH_SQL_QUERY;
-import static org.apache.ignite.internal.processors.odbc.OdbcRequest.GET_COLUMNS_META;
-import static org.apache.ignite.internal.processors.odbc.OdbcRequest.GET_TABLES_META;
 
 /**
  * SQL query handler.
  */
-public class OdbcNioListener extends GridNioServerListenerAdapter<OdbcRequest> {
-    /** Query ID sequence. */
-    private static final AtomicLong qryIdGen = new AtomicLong();
+public class OdbcNioListener extends GridNioServerListenerAdapter<byte[]> {
+    /** Initial output stream capacity. */
+    private static final int INIT_CAP = 1024;
 
     /** Kernel context. */
     private final GridKernalContext ctx;
@@ -63,17 +52,24 @@ public class OdbcNioListener extends GridNioServerListenerAdapter<OdbcRequest> {
     /** Logger. */
     private final IgniteLogger log;
 
-    /** Current queries cursors. */
-    private final ConcurrentHashMap<Long, IgniteBiTuple<QueryCursor, Iterator>> qryCurs = new ConcurrentHashMap<>();
+    /** Request handler. */
+    private final OdbcRequestHandler handler;
+
+    /** Marshaller. */
+    private final GridBinaryMarshaller marsh;
 
     /**
      * Constructor.
      *
      * @param ctx Context.
      */
-    public OdbcNioListener(final GridKernalContext ctx, final GridSpinBusyLock busyLock) {
+    public OdbcNioListener(final GridKernalContext ctx, final GridSpinBusyLock busyLock,
+                           final OdbcRequestHandler handler) {
         this.ctx = ctx;
         this.busyLock = busyLock;
+        this.handler = handler;
+
+        this.marsh = ((CacheObjectBinaryProcessorImpl)ctx.cacheObjects()).marshaller();
 
         this.log = ctx.log(getClass());
     }
@@ -98,18 +94,254 @@ public class OdbcNioListener extends GridNioServerListenerAdapter<OdbcRequest> {
     }
 
     /** {@inheritDoc} */
-    @Override public void onMessage(GridNioSession ses, OdbcRequest msg) {
+    @Override public void onMessage(GridNioSession ses, byte[] msg) {
         assert msg != null;
 
-        if (log.isDebugEnabled())
-            log.debug("Received request from client: [msg=" + msg + ']');
+        try {
+            OdbcRequest req = decode(msg);
 
-        OdbcResponse res = handle(msg);
+            OdbcResponse rsp = handle(req);
 
-        if (log.isDebugEnabled())
-            log.debug("Handling result: [res=" + res.status() + ']');
+            if (log.isDebugEnabled())
+                log.debug("Handling result: [res=" + rsp.status() + ']');
 
-        ses.send(res);
+            byte[] outMsg = encode(rsp);
+
+            ses.send(outMsg);
+        }
+        catch (Exception e) {
+            trySendErrorMessage(ses, e.getMessage());
+        }
+    }
+
+    /**
+     * Try to send simple response message to ODBC driver.
+     * @param ses Session.
+     * @param err Error message.
+     */
+    private void trySendErrorMessage(GridNioSession ses, String err) {
+        log.error(err);
+
+        try {
+            ses.send(encode(new OdbcResponse(OdbcResponse.STATUS_FAILED, err)));
+        }
+        catch (Exception e) {
+            log.error("Can not send error response message: [err=" + e.getMessage() + ']');
+        }
+    }
+
+    /**
+     * Decode OdbcRequest from byte array.
+     *
+     * @param msg Message.
+     * @return Assembled ODBC request.
+     */
+    private OdbcRequest decode(byte[] msg) throws IOException {
+        assert msg != null;
+
+        BinaryInputStream stream = new BinaryHeapInputStream(msg);
+
+        BinaryReaderExImpl reader = new BinaryReaderExImpl(null, stream, null);
+
+        OdbcRequest res;
+
+        byte cmd = reader.readByte();
+
+        switch (cmd) {
+            case OdbcRequest.EXECUTE_SQL_QUERY: {
+
+                String cache = reader.readString();
+                String sql = reader.readString();
+                int argsNum = reader.readInt();
+
+                if (log.isDebugEnabled()) {
+                    log.debug("Message: [cmd=EXECUTE_SQL_QUERY, cache=" + cache +
+                            ", query=" + sql + ", argsNum=" + argsNum + ']');
+                }
+
+                Object[] params = new Object[argsNum];
+
+                for (int i = 0; i < argsNum; ++i)
+                    params[i] = reader.readObjectDetached();
+
+                res = new OdbcQueryExecuteRequest(cache, sql, params);
+
+                break;
+            }
+
+            case OdbcRequest.FETCH_SQL_QUERY: {
+
+                long queryId = reader.readLong();
+                int pageSize = reader.readInt();
+
+                if (log.isDebugEnabled())
+                    log.debug("Message: [cmd=FETCH_SQL_QUERY, queryId=" + queryId + ", pageSize=" + pageSize + ']');
+
+                res = new OdbcQueryFetchRequest(queryId, pageSize);
+
+                break;
+            }
+
+            case OdbcRequest.CLOSE_SQL_QUERY: {
+
+                long queryId = reader.readLong();
+
+                if (log.isDebugEnabled()) {
+                    log.debug("Message: [cmd=CLOSE_SQL_QUERY, queryId=" + queryId + ']');
+                }
+
+                res = new OdbcQueryCloseRequest(queryId);
+
+                break;
+            }
+
+            case OdbcRequest.GET_COLUMNS_META: {
+
+                String cache = reader.readString();
+                String table = reader.readString();
+                String column = reader.readString();
+
+                if (log.isDebugEnabled()) {
+                    log.debug("Message: [cmd=GET_COLUMNS_META, cache=" + cache +
+                            ", table=" + table + ", column: " + column + ']');
+                }
+
+                res = new OdbcQueryGetColumnsMetaRequest(cache, table, column);
+
+                break;
+            }
+
+            case OdbcRequest.GET_TABLES_META: {
+
+                String catalog = reader.readString();
+                String schema = reader.readString();
+                String table = reader.readString();
+                String tableType = reader.readString();
+
+                if (log.isDebugEnabled()) {
+                    log.debug("Message: [cmd=GET_COLUMNS_META, catalog=" + catalog +
+                            ", schema=" + schema + ", table=" + table + ", tableType=" + tableType + ']');
+                }
+
+                res = new OdbcQueryGetTablesMetaRequest(catalog, schema, table, tableType);
+
+                break;
+            }
+
+            default:
+                throw new IOException("Failed to parse incoming packet (unknown command type) " +
+                        "[cmd=[" + Byte.toString(cmd) + ']');
+        }
+
+        return res;
+    }
+
+    /**
+     * Encode OdbcResponse to byte array.
+     *
+     * @param msg Message.
+     * @return Byte array.
+     */
+    private byte[] encode(OdbcResponse msg) throws IOException {
+        assert msg != null;
+
+        // Creating new binary writer
+        BinaryRawWriterEx writer = marsh.writer(new BinaryHeapOutputStream(INIT_CAP));
+
+        // Writing status
+        writer.writeByte((byte) msg.status());
+
+        if (msg.status() != OdbcResponse.STATUS_SUCCESS) {
+            writer.writeString(msg.error());
+
+            return Arrays.copyOf(writer.out().array(), writer.out().position());
+        }
+
+        Object res0 = msg.response();
+
+        if (res0 instanceof OdbcQueryExecuteResult) {
+            OdbcQueryExecuteResult res = (OdbcQueryExecuteResult) res0;
+
+            if (log.isDebugEnabled())
+                log.debug("Resulting query ID: " + res.getQueryId());
+
+            writer.writeLong(res.getQueryId());
+
+            Collection<OdbcColumnMeta> metas = res.getColumnsMetadata();
+
+            assert metas != null;
+
+            writer.writeInt(metas.size());
+
+            for (OdbcColumnMeta meta : metas)
+                meta.writeBinary(writer, marsh.context());
+
+        }
+        else if (res0 instanceof OdbcQueryFetchResult) {
+            OdbcQueryFetchResult res = (OdbcQueryFetchResult) res0;
+
+            if (log.isDebugEnabled())
+                log.debug("Resulting query ID: " + res.queryId());
+
+            writer.writeLong(res.queryId());
+
+            Collection<?> items0 = res.items();
+
+            assert items0 != null;
+
+            writer.writeBoolean(res.last());
+
+            writer.writeInt(items0.size());
+
+            for (Object row0 : items0) {
+                if (row0 != null) {
+                    Collection<?> row = (Collection<?>)row0;
+
+                    writer.writeInt(row.size());
+
+                    for (Object obj : row)
+                        writer.writeObjectDetached(obj);
+                }
+            }
+        }
+        else if (res0 instanceof OdbcQueryCloseResult) {
+            OdbcQueryCloseResult res = (OdbcQueryCloseResult) res0;
+
+            if (log.isDebugEnabled())
+                log.debug("Resulting query ID: " + res.getQueryId());
+
+            writer.writeLong(res.getQueryId());
+
+        }
+        else if (res0 instanceof OdbcQueryGetColumnsMetaResult) {
+            OdbcQueryGetColumnsMetaResult res = (OdbcQueryGetColumnsMetaResult) res0;
+
+            Collection<OdbcColumnMeta> columnsMeta = res.meta();
+
+            assert columnsMeta != null;
+
+            writer.writeInt(columnsMeta.size());
+
+            for (OdbcColumnMeta columnMeta : columnsMeta)
+                columnMeta.writeBinary(writer, marsh.context());
+
+        }
+        else if (res0 instanceof OdbcQueryGetTablesMetaResult) {
+            OdbcQueryGetTablesMetaResult res = (OdbcQueryGetTablesMetaResult) res0;
+
+            Collection<OdbcTableMeta> tablesMeta = res.meta();
+
+            assert tablesMeta != null;
+
+            writer.writeInt(tablesMeta.size());
+
+            for (OdbcTableMeta tableMeta : tablesMeta)
+                tableMeta.writeBinary(writer);
+        }
+        else
+            throw new IOException("Failed to serialize response packet (unknown response type)");
+
+        return Arrays.copyOf(writer.out().array(), writer.out().position());
     }
 
     /**
@@ -118,8 +350,11 @@ public class OdbcNioListener extends GridNioServerListenerAdapter<OdbcRequest> {
      * @param req Request.
      * @return Response.
      */
-    public OdbcResponse handle(OdbcRequest req) {
+    private OdbcResponse handle(OdbcRequest req) {
         assert req != null;
+
+        if (log.isDebugEnabled())
+            log.debug("Received request from client: [req=" + req + ']');
 
         if (!busyLock.enterBusy()) {
             String errMsg = "Failed to handle request [req=" + req +
@@ -131,270 +366,10 @@ public class OdbcNioListener extends GridNioServerListenerAdapter<OdbcRequest> {
         }
 
         try {
-            switch (req.command()) {
-                case EXECUTE_SQL_QUERY:
-                    return executeQuery((OdbcQueryExecuteRequest)req);
-
-                case FETCH_SQL_QUERY:
-                    return fetchQuery((OdbcQueryFetchRequest)req);
-
-                case CLOSE_SQL_QUERY:
-                    return closeQuery((OdbcQueryCloseRequest)req);
-
-                case GET_COLUMNS_META:
-                    return getColumnsMeta((OdbcQueryGetColumnsMetaRequest) req);
-
-                case GET_TABLES_META:
-                    return getTablesMeta((OdbcQueryGetTablesMetaRequest) req);
-            }
-
-            return new OdbcResponse(OdbcResponse.STATUS_FAILED,
-                    "Failed to find registered handler for command: " + req.command());
+            return handler.handle(req);
         }
         finally {
             busyLock.leaveBusy();
         }
-    }
-
-    /**
-     * {@link OdbcQueryExecuteRequest} command handler.
-     *
-     * @param req Execute query request.
-     * @return Response.
-     */
-    private OdbcResponse executeQuery(OdbcQueryExecuteRequest req) {
-        long qryId = qryIdGen.getAndIncrement();
-
-        try {
-            SqlFieldsQuery qry = new SqlFieldsQuery(req.sqlQuery());
-
-            qry.setArgs(req.arguments());
-
-            IgniteCache<Object, Object> cache = ctx.grid().cache(req.cacheName());
-
-            if (cache == null)
-                return new OdbcResponse(OdbcResponse.STATUS_FAILED,
-                    "Failed to find cache with name: " + req.cacheName());
-
-            QueryCursor qryCur = cache.query(qry);
-
-            Iterator cur = qryCur.iterator();
-
-            qryCurs.put(qryId, new IgniteBiTuple<>(qryCur, cur));
-
-            List<?> fieldsMeta = ((QueryCursorImpl) qryCur).fieldsMeta();
-
-            if (log.isDebugEnabled())
-                log.debug("Field meta: " + fieldsMeta);
-
-            OdbcQueryExecuteResult res = new OdbcQueryExecuteResult(qryId, convertMetadata(fieldsMeta));
-
-            return new OdbcResponse(res);
-        }
-        catch (Exception e) {
-            qryCurs.remove(qryId);
-
-            return new OdbcResponse(OdbcResponse.STATUS_FAILED, e.getMessage());
-        }
-    }
-
-    /**
-     * {@link OdbcQueryCloseRequest} command handler.
-     *
-     * @param req Execute query request.
-     * @return Response.
-     */
-    private OdbcResponse closeQuery(OdbcQueryCloseRequest req) {
-        try {
-            QueryCursor cur = qryCurs.get(req.queryId()).get1();
-
-            if (cur == null)
-                return new OdbcResponse(OdbcResponse.STATUS_FAILED, "Failed to find query with ID: " + req.queryId());
-
-            cur.close();
-
-            qryCurs.remove(req.queryId());
-
-            OdbcQueryCloseResult res = new OdbcQueryCloseResult(req.queryId());
-
-            return new OdbcResponse(res);
-        }
-        catch (Exception e) {
-            qryCurs.remove(req.queryId());
-
-            return new OdbcResponse(OdbcResponse.STATUS_FAILED, e.getMessage());
-        }
-    }
-
-    /**
-     * {@link OdbcQueryFetchRequest} command handler.
-     *
-     * @param req Execute query request.
-     * @return Response.
-     */
-    private OdbcResponse fetchQuery(OdbcQueryFetchRequest req) {
-        try {
-            Iterator cur = qryCurs.get(req.queryId()).get2();
-
-            if (cur == null)
-                return new OdbcResponse(OdbcResponse.STATUS_FAILED,
-                        "Failed to find query with ID: " + req.queryId());
-
-            List<Object> items = new ArrayList<>();
-
-            for (int i = 0; i < req.pageSize() && cur.hasNext(); ++i)
-                items.add(cur.next());
-
-            OdbcQueryFetchResult res = new OdbcQueryFetchResult(req.queryId(), items, !cur.hasNext());
-
-            return new OdbcResponse(res);
-        }
-        catch (Exception e) {
-            qryCurs.remove(req.queryId());
-
-            return new OdbcResponse(OdbcResponse.STATUS_FAILED, e.getMessage());
-        }
-    }
-
-    /**
-     * {@link OdbcQueryGetColumnsMetaRequest} command handler.
-     *
-     * @param req Get columns metadata request.
-     * @return Response.
-     */
-    private OdbcResponse getColumnsMeta(OdbcQueryGetColumnsMetaRequest req) {
-        try {
-            List<OdbcColumnMeta> meta = new ArrayList<>();
-
-            String cacheName;
-            String tableName;
-
-            if (req.tableName().contains(".")) {
-                // Parsing two-part table name.
-                String[] parts = req.tableName().split("\\.");
-
-                cacheName = removeQuotationMarksIfNeeded(parts[0]);
-
-                tableName = parts[1];
-            }
-            else {
-                cacheName = removeQuotationMarksIfNeeded(req.cacheName());
-
-                tableName = req.tableName();
-            }
-
-            Collection<GridQueryTypeDescriptor> tablesMeta = ctx.query().types(cacheName);
-
-            for (GridQueryTypeDescriptor table : tablesMeta) {
-                if (!matches(table.name(), tableName))
-                    continue;
-
-                for (Map.Entry<String, Class<?>> field : table.fields().entrySet()) {
-                    if (!matches(field.getKey(), req.columnName()))
-                        continue;
-
-                    OdbcColumnMeta columnMeta = new OdbcColumnMeta(req.cacheName(),
-                            table.name(), field.getKey(), field.getValue());
-
-                    if (!meta.contains(columnMeta))
-                        meta.add(columnMeta);
-                }
-            }
-            OdbcQueryGetColumnsMetaResult res = new OdbcQueryGetColumnsMetaResult(meta);
-
-            return new OdbcResponse(res);
-        }
-        catch (Exception e) {
-            return new OdbcResponse(OdbcResponse.STATUS_FAILED, e.getMessage());
-        }
-    }
-
-    /**
-     * {@link OdbcQueryGetTablesMetaRequest} command handler.
-     *
-     * @param req Get tables metadata request.
-     * @return Response.
-     */
-    private OdbcResponse getTablesMeta(OdbcQueryGetTablesMetaRequest req) {
-        try {
-            List<OdbcTableMeta> meta = new ArrayList<>();
-
-            String realSchema = removeQuotationMarksIfNeeded(req.schema());
-
-            for (String cacheName : ctx.cache().cacheNames())
-            {
-                if (!matches(cacheName, realSchema))
-                    continue;
-
-                Collection<GridQueryTypeDescriptor> tablesMeta = ctx.query().types(cacheName);
-
-                for (GridQueryTypeDescriptor table : tablesMeta) {
-                    if (!matches(table.name(), req.table()))
-                        continue;
-
-                    if (!matches("TABLE", req.tableType()))
-                        continue;
-
-                    OdbcTableMeta tableMeta = new OdbcTableMeta(req.catalog(), cacheName,
-                            table.name(), "TABLE");
-
-                    if (!meta.contains(tableMeta))
-                        meta.add(tableMeta);
-                }
-            }
-
-            OdbcQueryGetTablesMetaResult res = new OdbcQueryGetTablesMetaResult(meta);
-
-            return new OdbcResponse(res);
-        }
-        catch (Exception e) {
-            return new OdbcResponse(OdbcResponse.STATUS_FAILED, e.getMessage());
-        }
-    }
-
-    /**
-     * Convert metadata in collection from {@link GridQueryFieldMetadata} to
-     * {@link OdbcColumnMeta}.
-     *
-     * @param meta Internal query field metadata.
-     * @return Odbc query field metadata.
-     */
-    private static Collection<OdbcColumnMeta> convertMetadata(Collection<?> meta) {
-        List<OdbcColumnMeta> res = new ArrayList<>();
-
-        if (meta != null) {
-            for (Object info : meta) {
-                assert info instanceof GridQueryFieldMetadata;
-
-                res.add(new OdbcColumnMeta((GridQueryFieldMetadata)info));
-            }
-        }
-
-        return res;
-    }
-
-    /**
-     * Checks whether string matches SQL pattern.
-     *
-     * @param str String.
-     * @param ptrn Pattern.
-     * @return Whether string matches pattern.
-     */
-    private static boolean matches(String str, String ptrn) {
-        return str != null && (F.isEmpty(ptrn) ||
-            str.toUpperCase().matches(ptrn.toUpperCase().replace("%", ".*").replace("_", ".")));
-    }
-
-    /**
-     * Remove quotation marks at the beginning and end of the string if present.
-     *
-     * @param str Input string.
-     * @return String without leading and trailing quotation marks.
-     */
-    private static String removeQuotationMarksIfNeeded(String str) {
-        if (str.startsWith("\"") && str.endsWith("\""))
-            return str.substring(1, str.length() - 1);
-
-        return str;
     }
 }
